init and bot improvements
