+++ packages/common/src/constants/BotUserAgent.ts
@@ -1,0 +1,1 @@
+export const BotUserAgent = 'XREngine Puppeteer'
