+++ packages/engine/src/bot/functions/botHookFunctions.ts
@@ -2,7 +2,7 @@ import { MathUtils, Quaternion, Vector3 } from 'three'
 
 import { Engine } from '../../ecs/classes/Engine'
 import { getEngineState } from '../../ecs/classes/EngineState'
-import { getComponent } from '../../ecs/functions/ComponentFunctions'
+import { getComponent, hasComponent } from '../../ecs/functions/ComponentFunctions'
 import { useWorld } from '../../ecs/functions/SystemHooks'
 import { TransformComponent } from '../../transform/components/TransformComponent'
 import { BotHooks, XRBotHooks } from '../enums/BotHooks'
@@ -21,7 +21,6 @@ import {
 } from './xrBotHookFunctions'
 
 export const BotHookFunctions = {
-  [BotHooks.InitializeBot]: initializeBot,
   [BotHooks.LocationLoaded]: locationLoaded,
   [BotHooks.SceneLoaded]: sceneLoaded,
   [BotHooks.GetPlayerPosition]: getPlayerPosition,
@@ -41,10 +40,6 @@ export const BotHookFunctions = {
   [XRBotHooks.TweenXRInputSource]: tweenXRInputSource
 }
 
-export function initializeBot() {
-  Engine.instance.isBot = true
-}
-
 // === ENGINE === //
 
 export function locationLoaded() {
@@ -56,7 +51,12 @@ export function sceneLoaded() {
 }
 
 export function getPlayerPosition() {
-  return getComponent(useWorld().localClientEntity, TransformComponent)?.position
+  // console.log('get player position', Engine.instance.currentWorld.localClientEntity, hasComponent(Engine.instance.currentWorld.localClientEntity, TransformComponent))
+  const pos = getComponent(Engine.instance.currentWorld.localClientEntity, TransformComponent)?.position
+  // console.log(pos)
+  if (!pos) return
+  console.log(pos.x, pos.y, pos.z)
+  return getComponent(Engine.instance.currentWorld.localClientEntity, TransformComponent)?.position
 }
 
 export function getSceneMetadata() {
