+++ packages/engine/src/bot/systems/BotHookSystem.ts
@@ -4,9 +4,12 @@ import { AvatarInputSchema } from '../../avatar/AvatarInputSchema'
 import { Engine } from '../../ecs/classes/Engine'
 import { World } from '../../ecs/classes/World'
 import { EngineRenderer } from '../../renderer/WebGLRendererSystem'
+import { BotHookFunctions } from '../functions/botHookFunctions'
 import { sendXRInputData, simulateXR } from '../functions/xrBotHookFunctions'
 
 export default async function BotHookSystem(world: World) {
+  globalThis.botHooks = BotHookFunctions
+
   if (isDev) {
     const setupBotKey = 'setupBotKey'
     AvatarInputSchema.inputMap.set('Semicolon', setupBotKey)
